requirements.txt
aiofiles==23.2.1
aiogoogle==5.7.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.2.0
cachetools==5.3.3
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.51.0
frozenlist==1.4.1
google-auth==2.29.0
greenlet==3.0.3
idna==3.7
kiwisolver==1.4.5
loguru==0.7.2
Mako==1.3.3
MarkupSafe==2.1.5
matplotlib==3.9.0
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyaes==1.6.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.29
TgCrypto==1.2.5
tonyg-rfc3339==0.1
typing_extensions==4.11.0
tzdata==2024.1
win32-setctime==1.1.0
yarl==1.9.4
